Simplified travis and abs

+++ clr/hsv_test.go
@@ -19,10 +19,10 @@ func colorEquality(a RGB, b RGB) bool {
 	return abs(a.R, b.R) <= 1 && abs(a.G, b.G) <= 1 && abs(a.B, b.B) <= 1
 }
 
-func abs(a uint8, b uint8) uint8 {
+func abs(a uint8, b uint8) int {
 	c := int(a) - int(b)
 	if c < 0 {
 		c *= -1
 	}
-	return uint8(c)
+	return c
 }
